requirements.txt
aiosqlite==0.19.0
alembic==1.12.0
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
dnspython==2.4.2
email-validator==2.0.0.post2
fastadmin==0.1.41
fastapi==0.103.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
kombu==5.3.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
